% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allelecount.R
\name{dumpCounts.maf}
\alias{dumpCounts.maf}
\title{Run alleleCount
Dump allele counts from maf file}
\usage{
dumpCounts.maf(
  maf_file,
  tumour_outfile,
  normal_outfile = NA,
  samplename = NA,
  dummy_alt_allele = NA,
  dummy_ref_allele = NA
)
}
\arguments{
\item{maf_file}{The maf file to read in}

\item{tumour_outfile}{File to save the tumour counts to}

\item{normal_outfile}{Optional parameter specifying where the normal output should go}

\item{samplename}{Optional parameter specifying the samplename to be used for matching the right column in the maf}

\item{dummy_alt_allele}{}

\item{dummy_ref_allele}{}
}
\description{
Dump allele counts stored in the sample columns of the maf file. Output will go into a file
supplied as tumour_outfile and optionally normal_outfile. It will be a fully formatted
allele counts file as returned by alleleCounter.
}
\author{
sd11
}
