% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{maf2loci}
\alias{maf2loci}
\title{Transform vcf to loci file}
\usage{
maf2loci(maf_file, outfile, dummy_alt_allele = NA, dummy_ref_allele = NA)
}
\arguments{
\item{outfile}{Where to store the output}

\item{dummy_alt_allele}{The alt allele to store, supply when you want to override what is in the VCF (Default: NA)}

\item{dummy_ref_allele}{The reference allele to store, supply when you want to override what is in the VCF (Default: NA)}

\item{vcf_files}{A vector of vcf files to be considered}

\item{fai_file}{Reference genome index}

\item{ign_file}{A file with chromosomes to be excluded from consideration}
}
\description{
Function that dumps the loci of snvs from a series of vcf files into a single loci file
}
\author{
sd11
}
